self_extending_agent/requirements.txt
openai-agents>=0.0.1
python-dotenv>=0.20.0
pydantic>=2.0.0
